refactor(services): creates folders for every service
